feat(repos): one shared repo card with path labels and full action row
The repos grid, the home repos row, and the pinned repos in the projects
sidebar each had their own card markup, so they drifted: the grid and home
cards led with the badge short name, and only the sidebar card carried the
repo actions.

Collapse all three onto a single RepoCard:

- The card title is now the full repo/subpath, rendered by the shared
  RepoLabel (muted prefix, badge-hue emphasis on the distinguishing
  segment) instead of the short name. RepoLabel gains a `wrap` prop so a
  long path flows over as many lines as it needs rather than truncating.
- Under the title sits the action row previously exclusive to the sidebar:
  new project, run (cloned) or clone (not cloned), pin/unpin, and the more
  menu with "Open in…" and "Copy Path". Unpin moves out of the more menu
  into a one-click pin toggle on every surface, replacing the grid's
  floating corner toggle.
- The sidebar's pinned repos are the same card in `reorderable` mode, so
  they keep drag-to-reorder while picking up the grid's tint, border and
  layout.

The more menu now resolves the clone path and opener apps when it first
opens instead of on mount, so a grid of N repos no longer fires N
get_repo_clone_path calls up front.

SidebarPinnedRepo is deleted, and the grid's duplicated card CSS along with
it; ReposListView and ProjectsList only own their own layout (grid cell
stretch, 200px scroll-row width).

Verified with `just typecheck`, `just fmt-check`, `just test-frontend`, and
a `VITE_REPOS_UI_ENABLED=true` production build. The repos UI stays behind
that flag, so this is dark by default.
